i changed ERXGenericRecord and now it uses reflection to call Validity. Just set er.extensions.ERXGenericRecord.useValidity=true
in any Properties file that gets loaded or do this manually: System.setProperty("er.extensions.ERXGenericRecord.useValidity", "true");

and ERXGenericRecord will use validity. Make sure that validity classes are in the classpath by adding the framework, this is -not- necessary for compiling ERExtensions, its just important at runtime so that reflection finds the classes.


git-svn-id: file:///Users/q/development/git/svnroot/trunk/Wonder@1566 8dc76679-1c23-de11-85cd-0050568e728b