refinded error handling a bit: when you set the option to "create classes for EOs" in EOModeler, you will get distinct names for each class.
Unfortunately, when you don't want to create those classes now, you will get an error message.

This fix just warn you and set the class to EOGenericRecord. We might consider to wrap this behaviour into yet another conditional setting (ERXEntityClassDescription.shouldCopeWithMissingClass) if it poses a problem.

Also, we should make the class to which to switch configurable (ERXGenericRecord would be great)


git-svn-id: https://wonder.svn.sourceforge.net/svnroot/wonder/trunk/Wonder@1741 8dc76679-1c23-de11-85cd-0050568e728b