changed handling of empty strings and default values: the default with booleans is now the provided value, and empty strings when parsed as number are regarded as NULL.
This is because the D2W RuleEditor can no longer save NULL values but provides empty strings.

NOTE: this is a rather fundamental change and I have not really tested it with a lot of code - so if your code depends on "" being a number (which throws a NumberFormatException) then be sure to drp me a note.


git-svn-id: file:///Users/q/development/git/svnroot/trunk/Wonder@2083 8dc76679-1c23-de11-85cd-0050568e728b