I removed the need for a wrapping D2WCustomComponent a while ago so we can use ERDXXXCustomComponents in the D2WAssistant.
However, with some components (ERDEditDateJavascript for example), we will need to pull the name of the key from the "d2wContext.propertyKey" instead of "key" because the component is synching, but does not have a "key" binding set.

I hope this is OK - with my stuff it is...but I'd prefer to have more comments on this.


git-svn-id: file:///Users/q/development/git/svnroot/trunk@2085 8dc76679-1c23-de11-85cd-0050568e728b