patching head with change from branch:
Changes to support a willRevert/didRevert message to the EC delegate and message
to the reverting EOs.

1/ added a ERXEC.Delegate interface that extends the EOEditingContext.Delegate
interface which is what we used to call an informal protocol.  this delegate
interface adds two methods for before and after the revert.
2/ changed ERXEC.revert() method to invoke the delegate methods if they're implemented
by the delegate.
3/ added delegate implementations on ERXDefaultEditingContextDelegate to call
willRevert() and didRevert() for any ERXGenericRecord subclasses.
4/ implemented willRevert() and didRevert() on ERXGenericRecord

ERXGenericRecord's willRevert does nothing, didRevert flushes caches.


git-svn-id: file:///Users/q/development/git/svnroot/trunk/Wonder@3556 8dc76679-1c23-de11-85cd-0050568e728b