Chris Meyer: This patch modifies the EROpenID framework to handle different kinds of OpenID message extensions. This allows clients to (potentially) handle "simple registration" messages as well as "attribute exchange" extensions. The old implementation only allowed for the newer "attribute exchange" extensions; and only one of those. The new implementation allows for more message extension types and now multiple ones can be requested too.
This patch is in preparation for additional patches to make it easy to retrieve additional registration information (such as an email address) without much effort on the client side.

This patch should be 100% backwards compatible although I have marked the older API as deprecated. I strongly doubt that anyone was using the older code since it was so limited in functionality, but any feedback would be welcome, of course.

This patch also adds several logging messages.


git-svn-id: https://wonder.svn.sourceforge.net/svnroot/wonder/trunk/Wonder@6906 8dc76679-1c23-de11-85cd-0050568e728b