Log warnings when accessing certain values in ERXThreadStorage that were inherited from the parent thread.
Found two places in Wonder so far, where this was tripped:

One was ERXEOAccessUtilities.modelGroup(ec). This is now checking if the current thread storage map was inherited and falls back to use EOEmodelGroup.defaultGroup() if this is the case. 

The other place was ERXObjectStoreCoordinatorPool which accessed the inherited session to get the sessionID. This now trips a warning, although in this special case the usage is unproblematic, as only "safe" parts of the session are accessed. To get around the warning, ERXSession now also stores it's sessionID separately in the thread storage, so it can be accessed without touching the session object itself. 



git-svn-id: https://wonder.svn.sourceforge.net/svnroot/wonder/trunk/Wonder@8715 8dc76679-1c23-de11-85cd-0050568e728b