Patch from Mark Ritchie.
Fixed the case of the Apache conditional so that serious error are actually reported to Apache.  Changed the code to match the current Apache API.
Removed an Apache conditional which was never true because the case was wrong.
Log a message to the apache log when we fail to write messages as directed by WebObjectsLog configuration.
Changed to work with clients which quote woinst cookie values and added message when we fail to find a matching instance.



git-svn-id: https://wonder.svn.sourceforge.net/svnroot/wonder/trunk@10806 8dc76679-1c23-de11-85cd-0050568e728b