http://issues.objectstyle.org/jira/browse/WONDER-552
[PATCH] Fix to make the method really work with keyPaths

The method er.extensions.eof.ERXEOControlUtilities.distinctValuesForKeyPath(EOEditingContext, String, String, EOQualifier, NSArray<EOSortOrdering>) current works for cases when the keyPath is just a key of the entityName parameter. However it throws an error when a keyPath to a related object is passed in. Also when a keyPath to related objects is passed in, the result will have duplicates. This patch fixes these two problems encountered when using this method with keyPaths.

git-svn-id: https://wonder.svn.sourceforge.net/svnroot/wonder/trunk@10988 8dc76679-1c23-de11-85cd-0050568e728b